blacklist implimented - needs testing

+++ src/chaptersToDownload.h
@@ -17,8 +17,8 @@ typedef struct chapterqueue ChapterQueue;
 
 void set_series_folder(char *folder);
 void set_source(Site domainUsed);
-int get_download_length();
-int get_current_download_chapter();
+unsigned long get_download_length();
+unsigned long get_current_download_chapter();
 void add_to_download_list(Chapter *toAdd);
 void download_entire_queue();
 char *get_series_folder();
